factor out validators
