+++ apps/web/src/components/UserProfile/ProfileSections/AboutSection.tsx
@@ -12,58 +12,11 @@ import {
 
 import { getFullNameHtml } from "~/utils/nameUtils";
 import type { Applicant } from "~/api/generated";
-
-type PartialApplicant = Pick<
-  Applicant,
-  | "firstName"
-  | "lastName"
-  | "email"
-  | "telephone"
-  | "preferredLang"
-  | "preferredLanguageForInterview"
-  | "preferredLanguageForExam"
-  | "currentCity"
-  | "currentProvince"
-  | "citizenship"
-  | "armedForcesStatus"
->;
+import { hasEmptyRequiredFields } from "~/validators/profile/about";
 
 interface AboutSectionProps {
-  applicant: PartialApplicant;
+  applicant: Applicant;
   editPath?: string;
-}
-
-export function hasEmptyRequiredFields({
-  firstName,
-  lastName,
-  telephone,
-  email,
-  preferredLang,
-  preferredLanguageForInterview,
-  currentCity,
-  currentProvince,
-  citizenship,
-  armedForcesStatus,
-}: PartialApplicant): boolean {
-  return (
-    !firstName ||
-    !lastName ||
-    !email ||
-    !telephone ||
-    !preferredLang ||
-    !preferredLanguageForInterview ||
-    !preferredLanguageForInterview ||
-    !currentCity ||
-    !currentProvince ||
-    !citizenship ||
-    armedForcesStatus === null
-  );
-}
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export function hasEmptyOptionalFields(applicant: PartialApplicant): boolean {
-  // no optional fields
-  return false;
 }
 
 const AboutSection = ({ editPath, applicant }: AboutSectionProps) => {
