+++ apps/web/src/components/UserProfile/ProfileSections/LanguageInformationSection.tsx
@@ -9,67 +9,16 @@ import {
 } from "@gc-digital-talent/i18n";
 
 import { Applicant, BilingualEvaluation } from "~/api/generated";
-
-type PartialApplicant = Pick<
-  Applicant,
-  | "lookingForEnglish"
-  | "lookingForFrench"
-  | "lookingForBilingual"
-  | "bilingualEvaluation"
-  | "estimatedLanguageAbility"
-  | "writtenLevel"
-  | "comprehensionLevel"
-  | "verbalLevel"
->;
-
-export function hasAllEmptyFields({
-  lookingForEnglish,
-  lookingForFrench,
-  lookingForBilingual,
-  bilingualEvaluation,
-}: PartialApplicant): boolean {
-  return (
-    !lookingForEnglish &&
-    !lookingForFrench &&
-    !lookingForBilingual &&
-    !bilingualEvaluation
-  );
-}
-
-export function hasEmptyRequiredFields({
-  lookingForEnglish,
-  lookingForFrench,
-  lookingForBilingual,
-  bilingualEvaluation,
-  writtenLevel,
-  comprehensionLevel,
-  verbalLevel,
-}: PartialApplicant): boolean {
-  return !!(
-    (!lookingForEnglish && !lookingForFrench && !lookingForBilingual) ||
-    (lookingForBilingual &&
-      (!bilingualEvaluation ||
-        ((bilingualEvaluation === BilingualEvaluation.CompletedEnglish ||
-          bilingualEvaluation === BilingualEvaluation.CompletedFrench) &&
-          (!comprehensionLevel || !writtenLevel || !verbalLevel))))
-  );
-}
-
-export function hasEmptyOptionalFields({
-  bilingualEvaluation,
-  estimatedLanguageAbility,
-}: PartialApplicant): boolean {
-  return (
-    bilingualEvaluation === BilingualEvaluation.NotCompleted &&
-    !estimatedLanguageAbility
-  );
-}
+import {
+  hasAllEmptyFields,
+  hasEmptyRequiredFields,
+} from "~/validators/profile/languageInformation";
 
 const LanguageInformationSection = ({
   applicant,
   editPath,
 }: {
-  applicant: PartialApplicant;
+  applicant: Applicant;
   editPath?: string;
 }) => {
   const intl = useIntl();
