+++ apps/web/src/components/UserProfile/ProfileSections/WorkPreferencesSection.tsx
@@ -13,35 +13,16 @@ import {
   OperationalRequirement,
   PositionDuration,
 } from "~/api/generated";
-
-type PartialApplicant = Pick<
-  Applicant,
-  "acceptedOperationalRequirements" | "positionDuration"
->;
-
-export function hasAllEmptyFields({
-  positionDuration,
-}: PartialApplicant): boolean {
-  return isEmpty(positionDuration);
-}
-
-export function hasEmptyRequiredFields({
-  positionDuration,
-}: PartialApplicant): boolean {
-  return isEmpty(positionDuration);
-}
-
-export function hasEmptyOptionalFields({
-  acceptedOperationalRequirements,
-}: PartialApplicant): boolean {
-  return isEmpty(acceptedOperationalRequirements);
-}
+import {
+  hasAllEmptyFields,
+  hasEmptyRequiredFields,
+} from "~/validators/profile/workPreferences";
 
 const WorkPreferencesSection = ({
   applicant,
   editPath,
 }: {
-  applicant: PartialApplicant;
+  applicant: Applicant;
   editPath?: string;
 }) => {
   const intl = useIntl();
