+++ apps/web/src/components/UserProfile/UserProfile.tsx
@@ -18,38 +18,33 @@ import { useFeatureFlags } from "@gc-digital-talent/env";
 
 import type { Applicant } from "~/api/generated";
 
-import AboutSection, {
-  hasEmptyRequiredFields as aboutSectionHasEmptyRequiredFields,
-  hasEmptyOptionalFields as aboutSectionHasEmptyOptionalFields,
-} from "~/components/UserProfile/ProfileSections/AboutSection";
-import LanguageInformationSection, {
-  hasEmptyRequiredFields as languageInformationSectionHasEmptyRequiredFields,
-  hasEmptyOptionalFields as languageInformationSectionHasEmptyOptionalFields,
-} from "~/components/UserProfile/ProfileSections/LanguageInformationSection";
-import GovernmentInformationSection, {
-  hasEmptyRequiredFields as governmentInformationSectionHasEmptyRequiredFields,
-  hasEmptyOptionalFields as governmentInformationSectionHasEmptyOptionalFields,
-} from "~/components/UserProfile/ProfileSections/GovernmentInformationSection";
-import WorkLocationSection, {
-  hasEmptyRequiredFields as workLocationSectionHasEmptyRequiredFields,
-  hasEmptyOptionalFields as workLocationSectionHasEmptyOptionalFields,
-} from "~/components/UserProfile/ProfileSections/WorkLocationSection";
-import WorkPreferencesSection, {
-  hasEmptyRequiredFields as workPreferencesSectionHasEmptyRequiredFields,
-  hasEmptyOptionalFields as workPreferencesSectionHasEmptyOptionalFields,
-} from "~/components/UserProfile/ProfileSections/WorkPreferencesSection";
-import DiversityEquityInclusionSection, {
-  hasEmptyRequiredFields as diversityEquityInclusionSectionHasEmptyRequiredFields,
-  hasEmptyOptionalFields as diversityEquityInclusionSectionHasEmptyOptionalFields,
-} from "~/components/UserProfile/ProfileSections/DiversityEquityInclusionSection";
-import RoleSalarySection, {
-  hasEmptyRequiredFields as roleSalarySectionHasEmptyRequiredFields,
-  hasEmptyOptionalFields as roleSalarySectionHasEmptyOptionalFields,
-} from "~/components/UserProfile/ProfileSections/RoleSalarySection";
+import {
+  aboutSectionHasEmptyRequiredFields,
+  aboutSectionHasEmptyOptionalFields,
+  diversityEquityInclusionSectionHasEmptyRequiredFields,
+  diversityEquityInclusionSectionHasEmptyOptionalFields,
+  governmentInformationSectionHasEmptyRequiredFields,
+  governmentInformationSectionHasEmptyOptionalFields,
+  languageInformationSectionHasEmptyRequiredFields,
+  languageInformationSectionHasEmptyOptionalFields,
+  roleSalarySectionHasEmptyRequiredFields,
+  roleSalarySectionHasEmptyOptionalFields,
+  workLocationSectionHasEmptyRequiredFields,
+  workLocationSectionHasEmptyOptionalFields,
+  workPreferencesSectionHasEmptyRequiredFields,
+  workPreferencesSectionHasEmptyOptionalFields,
+} from "~/validators/profile";
 
 import ExperienceSection from "./ExperienceSection";
 import { StatusItem } from "../InfoItem";
 import { Status } from "../InfoItem/StatusItem";
+import AboutSection from "./ProfileSections/AboutSection";
+import DiversityEquityInclusionSection from "./ProfileSections/DiversityEquityInclusionSection";
+import GovernmentInformationSection from "./ProfileSections/GovernmentInformationSection";
+import LanguageInformationSection from "./ProfileSections/LanguageInformationSection";
+import RoleSalarySection from "./ProfileSections/RoleSalarySection";
+import WorkLocationSection from "./ProfileSections/WorkLocationSection";
+import WorkPreferencesSection from "./ProfileSections/WorkPreferencesSection";
 
 interface SectionControl {
   isVisible: boolean;
